+++ src/main/java/cn/gdufe/web/servlet/course/AddClassServlet.java
@@ -1,4 +1,4 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.course;
 
 import cn.gdufe.domain.Class;
 import cn.gdufe.service.ClassService;
