+++ src/main/java/cn/gdufe/web/servlet/student/DelStudentServlet.java
@@ -1,4 +1,4 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.student;
 
 import cn.gdufe.service.StudentService;
 import cn.gdufe.service.impl.StudentServiceImpl;
