+++ src/main/java/cn/gdufe/web/servlet/teacher/DelTeacherServlet.java
@@ -1,4 +1,4 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.teacher;
 
 import cn.gdufe.service.TeacherService;
 import cn.gdufe.service.impl.TeacherServiceImpl;
