+++ src/main/java/cn/gdufe/web/servlet/teacher/FindTeacherServlet.java
@@ -1,8 +1,9 @@
-package cn.gdufe.web.servlet;
+package cn.gdufe.web.servlet.teacher;
 
-import cn.gdufe.domain.Class;
-import cn.gdufe.service.ClassService;
-import cn.gdufe.service.impl.ClassServiceImpl;
+import cn.gdufe.domain.Teacher;
+import cn.gdufe.service.TeacherService;
+import cn.gdufe.service.impl.TeacherServiceImpl;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -10,23 +11,23 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/findClassServlet")
-public class FindClassServlet extends HttpServlet {
+@WebServlet("/findTeacherServlet")
+public class FindTeacherServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         //1.设置编码
         request.setCharacterEncoding("utf-8");
 
         //2.获取参数
-        String classId = request.getParameter("classId");
-        ClassService service = new ClassServiceImpl();
-        Class c = service.findClass(classId);
+        String tid = request.getParameter("tid");
+        TeacherService service = new TeacherServiceImpl();
+        Teacher teacher = service.findTeacher(tid);
 
         //3.将user存入request
-        request.setAttribute("c",c);
+        request.setAttribute("teacher",teacher);
 
         //4.转发到update.jsp
-        request.getRequestDispatcher("/edit_class.jsp").forward(request,response);
+        request.getRequestDispatcher("/edit_teacher.jsp").forward(request,response);
     }
 
     @Override
